pyproject.toml
[project]
name = "difftools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "optuna>=3.6.1",
    "pyyaml>=6.0.3",
    "rich>=13.9.4",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["model_runner/src"]
